pyproject.toml
[build-system]
requires = ["flit_core ~= 3.5"]
build-backend = "flit_core.buildapi"

[project]
name = "meshgraphnet"
dynamic = ["version", "description"]
readme = "README.md"
requires-python = ">= 3.9"
license = {file = "LICENSE"}
authors = [{name = "Marten Lienen", email = "marten.lienen@in.tum.de"}]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python"
]
dependencies = [
  "numpy ~= 1.19",
  "scipy ~= 1.7",
  "matplotlib ~= 3.4",
  "scikit-spatial ~= 6.2",
  "scikit-fem ~= 5.2",
  "einops ~= 0.4",
  "torch == 1.10",
  "torchdiffeq ~= 0.2.2",
  "torch-scatter ~= 2.0",
  "torch-sparse ~= 0.6",
  "torch-geometric ~= 2.0",
  "torchtyping",
  "tensorflow ~= 2.0",
  "more-itertools ~= 8.12",
  "cachetools ~= 5.0",
  "hydra-core ~= 1.1",
  "hydra-experiment-sweeper ~= 0.1",
  "hydra-submitit-launcher ~= 1.1",
  "omegaconf ~= 2.1",
  "wandb ~= 0.12"
]

[project.optional-dependencies]
lightning = [
  "pytorch-lightning ~= 1.5",
  "torchmetrics",
  "tqdm",
  "requests",
  "scikit-learn-extra"
]

[tool.flit.module]
name = "meshgraphnet"

[tool.black]
line-length = 88
target-version = ["py39"]
